Fix v6 Router Errors
