+++ frontend/src/index.js
@@ -22,6 +22,15 @@ import ReactDOM from "react-dom/client";
 
 import { BrowserRouter, Routes, Route, Navigate, Link, } from "react-router-dom";
 
+import routes from "routes.js";
+import Dashboard from "views/Dashboard.js";
+import UserProfile from "views/UserProfile.js";
+import TableList from "views/TableList.js";
+import Typography from "views/Typography.js";
+import Icons from "views/Icons.js";
+import Maps from "views/Maps.js";
+import Upgrade from "views/Upgrade.js";
+
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./assets/css/animate.min.css";
 import "./assets/scss/light-bootstrap-dashboard-react.scss?v=2.0.0";
@@ -32,11 +41,29 @@ import AdminLayout from "layouts/Admin.js";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
+const getRoutes = (routes) => {
+  return routes.map((prop, key) => {
+    if (prop.layout === "/admin") {
+      return (
+        <Route
+          path={prop.layout + prop.path}
+          element={<prop.component />}
+          key={key}
+        />
+      );
+    } else {
+      return <Route />;
+    }
+  });
+};
+
 root.render(
   <BrowserRouter>
     <Routes>
-      <Route path="/admin/*" element={<AdminLayout />} />
-      <Route path="*" element={<Navigate to="/admin/dashboard" replace />} />
+      <Route path="/" element={<Navigate to="/admin/dashboard" replace />} />
+      <Route element={<AdminLayout />}>
+        {getRoutes(routes)}
+      </Route>
     </Routes>
   </BrowserRouter>
 );
