+++ frontend/src/routes.js
@@ -1,3 +1,4 @@
+/* eslint-disable */ 
 /*!
 
 =========================================================
